backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.15.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
redis[hiredis]>=5.0.0
pykrx>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
apscheduler>=3.10.0
vectorbt>=0.26.0
python-telegram-bot>=22.0
python-json-logger>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 테스트 의존성
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.28.0
aiosqlite>=0.20.0
